ml_project/requirements.txt
antlr4-python3-runtime==4.9.3
contourpy==1.0.5
cycler==0.11.0
fonttools==4.38.0
hydra-core==1.2.0
importlib-resources==5.10.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.0
numpy==1.23.4
omegaconf==2.2.3
packaging==21.3
pandas==1.5.1
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sweetviz==2.1.4
threadpoolctl==3.1.0
tqdm==4.64.1
seaborn==0.12.1
black[jupyter]
flake8
click
loguru==0.6.0
faker==15.1.1